feat: RankingBook 연결 해제 및 Book과 Review 이용해서 queryDSL 재구성

+++ src/main/java/com/sprint/deokhugamteam7/domain/book/dto/BookActivity.java
@@ -4,7 +4,7 @@ import java.util.UUID;
 
 public record BookActivity(
     UUID bookId,
-    int reviewCount,
+    long reviewCount,
     int totalRating
 ) {
 }
